server_study/django_app/requirements.txt
django==2.1.10
pymysql==0.9.3
djangorestframework==3.10.1